Changed: Use explicit casts for places where the value is changed, presumably to avoid some compiler warnings.

+++ libcmyth/socket.c
@@ -102,7 +102,7 @@ cmyth_send_message(cmyth_conn_t conn, char *request)
 		tv.tv_usec = 0;
 		FD_ZERO(&fds);
 		FD_SET(conn->conn_fd, &fds);
-		if (select(conn->conn_fd+1, NULL, &fds, NULL, &tv) == 0) {
+		if (select((int)conn->conn_fd+1, NULL, &fds, NULL, &tv) == 0) {
 			conn->conn_hang = 1;
 			continue;
 		} else {
@@ -165,7 +165,7 @@ cmyth_rcv_length(cmyth_conn_t conn)
 		tv.tv_usec = 0;
 		FD_ZERO(&fds);
 		FD_SET(conn->conn_fd, &fds);
-		if ((r=select(conn->conn_fd+1, &fds, NULL, NULL, &tv)) == 0) {
+		if ((r=select((int)conn->conn_fd+1, &fds, NULL, NULL, &tv)) == 0) {
 			conn->conn_hang = 1;
 			continue;
 		} else if (r > 0) {
@@ -234,7 +234,7 @@ cmyth_conn_refill(cmyth_conn_t conn, int len)
 		tv.tv_usec = 0;
 		FD_ZERO(&fds);
 		FD_SET(conn->conn_fd, &fds);
-		if ((r=select(conn->conn_fd+1, &fds, NULL, NULL, &tv)) == 0) {
+		if ((r=select((int)conn->conn_fd+1, &fds, NULL, NULL, &tv)) == 0) {
 			conn->conn_hang = 1;
 			continue;
 		} else if (r > 0) {
@@ -570,7 +570,7 @@ cmyth_rcv_long(cmyth_conn_t conn, int *err, long *buf, int count)
 		 * Check and make sure we are still under the limit (this is
 		 * an absolute value limit, sign will be applied later).
 		 */
-		if (val > limit) {
+		if (val > (unsigned long)limit) {
 			cmyth_dbg(CMYTH_DBG_ERROR,
 				  "%s: long out of range: '%s'\n",
 				  __FUNCTION__, num);
@@ -2697,7 +2697,7 @@ cmyth_rcv_data(cmyth_conn_t conn, int *err, unsigned char *buf, int count)
 		tv.tv_usec = 0;
 		FD_ZERO(&fds);
 		FD_SET(conn->conn_fd, &fds);
-		if (select(conn->conn_fd+1, &fds, NULL, NULL, &tv) == 0) {
+		if (select((int)conn->conn_fd+1, &fds, NULL, NULL, &tv) == 0) {
 			conn->conn_hang = 1;
 			continue;
 		} else {
